Added endpoint to get set inventory!

+++ bricklink-api/src/main/java/io/github/followsclosely/bricklink/dto/BlinkItem.java
@@ -175,7 +175,7 @@ public class BlinkItem {
             @JsonProperty("item")
             private InventoryItem item;
             @JsonProperty("color_id")
-            private Integer colorId;
+            private Long colorId;
             @JsonProperty("quantity")
             private Integer quantity;
             @JsonProperty("extra_quantity")
